+++ core/tracky-mouse.js
@@ -1488,6 +1488,17 @@ You may want to turn this off if you're drawing on a canvas, or increase it if y
 
 	paused = !s.startEnabled;
 
+	// Basically Promise.withResolvers (but I'm not sure browser support is good enough)
+	function createDeferred() {
+		let resolve, reject;
+		const promise = new Promise((res, rej) => {
+			resolve = res;
+			reject = rej;
+		});
+		return { promise, resolve, reject };
+	}
+
+	let matchedCameraIdDeferred = createDeferred();
 	let populateCameraList = () => { };
 	if (navigator.mediaDevices && navigator.mediaDevices.enumerateDevices) {
 		populateCameraList = () => {
@@ -1524,27 +1535,35 @@ You may want to turn this off if you're drawing on a canvas, or increase it if y
 				defaultOption.text = "Default";
 				cameraSelect.appendChild(defaultOption);
 
-				let found = false;
+				let matchingDeviceId = "";
 				for (const device of videoDevices) {
 					const option = document.createElement('option');
 					option.value = device.deviceId;
 					option.text = device.label || `Camera ${cameraSelect.length}`;
 					cameraSelect.appendChild(option);
 					if (device.deviceId === s.cameraDeviceId) {
-						found = true;
+						matchingDeviceId = device.deviceId;
+					} else if (device.label === knownCameras[s.cameraDeviceId]?.name) {
+						matchingDeviceId ||= device.deviceId;
 					}
 				}
-				// Defaulting to "Default" would imply a preference isn't stored.
+
+				// Defaulting to "Default" would imply a preference isn't stored...
+				// but would it be more friendly anyways?
 				// cameraSelect.value = found ? s.cameraDeviceId : "";
+
 				// Show a placeholder for the selected camera
-				if (s.cameraDeviceId && !found) {
+				if (s.cameraDeviceId && !matchingDeviceId) {
 					const option = document.createElement("option");
 					option.value = s.cameraDeviceId;
 					const knownInfo = knownCameras[s.cameraDeviceId];
 					option.text = knownInfo ? `${knownInfo.name} (Unavailable)` : "Unavailable camera";
 					cameraSelect.appendChild(option);
+					cameraSelect.value = s.cameraDeviceId;
+				} else {
+					cameraSelect.value = matchingDeviceId;
 				}
-				cameraSelect.value = s.cameraDeviceId;
+				matchedCameraIdDeferred.resolve(matchingDeviceId);
 			});
 		};
 		populateCameraList();
@@ -1599,7 +1618,7 @@ You may want to turn this off if you're drawing on a canvas, or increase it if y
 		updateStartStopButton();
 	};
 
-	useCameraButton.onclick = TrackyMouse.useCamera = async () => {
+	useCameraButton.onclick = TrackyMouse.useCamera = async (optionsOrEvent) => {
 		await settingsLoadedPromise;
 		const constraints = {
 			audio: false,
@@ -1609,9 +1628,10 @@ You may want to turn this off if you're drawing on a canvas, or increase it if y
 				facingMode: "user",
 			}
 		};
-		if (s.cameraDeviceId) {
+		const deviceIdToTry = optionsOrEvent?.retryWithCameraDeviceId ?? s.cameraDeviceId;
+		if (deviceIdToTry) {
 			delete constraints.video.facingMode;
-			constraints.video.deviceId = { exact: s.cameraDeviceId };
+			constraints.video.deviceId = { exact: deviceIdToTry };
 		}
 		navigator.mediaDevices.getUserMedia(constraints).then((stream) => {
 			populateCameraList();
@@ -1620,7 +1640,50 @@ You may want to turn this off if you're drawing on a canvas, or increase it if y
 			cameraVideo.srcObject = stream;
 			useCameraButton.hidden = true;
 			errorMessage.hidden = true;
-		}, (error) => {
+		}, async (error) => {
+
+			// OverconstrainedError can be caused by `deviceId` not matching,
+			// either due to the device not being present, or the ID having changed (don't ask me why that can happen but it can)
+			// Note: OverconstrainedError has a `constraint` property but not in Firefox so it's not very helpful.
+			// Note: (Not sure about ConstraintNotSatisfiedError here)
+			if (
+				(error.name === "OverconstrainedError" || error.name == "ConstraintNotSatisfiedError") &&
+				constraints.video.deviceId?.exact
+			) {
+				// This is giving me User Gesture Hinged Access And Access Attempt Absorption Anxiety,
+				// or "UGHAaAAAA" (I'm coining that term)
+				const matchedCameraId = await matchedCameraIdDeferred.promise;
+				if (matchedCameraId) {
+					// TODO: make sure matchedCameraId !== deviceIdToTry
+					TrackyMouse.useCamera({ retryWithCameraDeviceId: matchedCameraId });
+				} else {
+					// TODO: unify code branches for error handling
+
+					// TODO: handle case where permission is no longer granted,
+					// and enumerateDevices returns a fake list
+					// and getUserMedia fails with OverconstrainedError when passed a real deviceId because it's not in the fake list
+					// It's possible we could connect to the device without the user having to change the device
+					// in the dropdown (twice, in case they want a non-default camera)
+					// by first calling getUserMedia with no deviceId constraint, then closing the stream,
+					// then enumerating devices (and updating the dropdown with the real info)
+					// and calling getUserMedia again with the deviceId from the settings.
+					// The user should only need to respond to a permissions prompt once.
+
+					console.error(error, { matchedCameraId, "s.cameraDeviceId": s.cameraDeviceId, knownCameras: JSON.parse(localStorage.getItem("tracky-mouse-known-cameras") || "{}"), videoDevices: await navigator.mediaDevices.enumerateDevices() });
+					// errorMessage.textContent = "The previously selected camera is not available. Please select a different camera from the dropdown and try again.";
+					// errorMessage.textContent = "The previously selected camera is not available. Please mess around with Video > Camera source.";
+					// errorMessage.textContent = "The previously selected camera is not available. Try changing Video > Camera source.";
+					// errorMessage.textContent = "The previously selected camera is not available. Please select a camera from the \"Camera source\" dropdown in the Video settings and if it doesn't show up, it might after you select Default.";
+					errorMessage.textContent = "The previously selected camera is not available. Try selecting \"Default\" for Video > Camera source, and then select a specific camera if you need to.";
+					// It's awkward but that's my best attempt at conveying how you may need to proceed
+					// without complicated description of how/why the dropdown might be populated with
+					// fake information until a camera stream is successfully opened.
+					errorMessage.textContent = `⚠️ ${errorMessage.textContent}`;
+					errorMessage.hidden = false;
+				}
+				return;
+			}
+
 			console.log(error);
 			if (error.name == "NotFoundError" || error.name == "DevicesNotFoundError") {
 				// required track is missing
@@ -1639,9 +1702,7 @@ You may want to turn this off if you're drawing on a canvas, or increase it if y
 				errorMessage.textContent = "Webcam may already be in use. Please make sure you have no other programs using the camera.";
 			} else if (error.name == "OverconstrainedError" || error.name == "ConstraintNotSatisfiedError") {
 				// constraints cannot be satisfied by available devices
-				// TODO: improve handling of OverconstrainedError
-				// now that we constrain on a deviceId if a camera has been selected
-				// https://github.com/1j01/tracky-mouse/issues/122
+				// NOTE: handled above
 				errorMessage.textContent = `Webcam does not support the required resolution. Please change your settings.`;
 			} else if (error.name == "NotAllowedError" || error.name == "PermissionDeniedError") {
 				// permission denied in browser